html/rendering/bindings/the-marquee-element/marquee_scrolldelay.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: marquee_scrolldelay</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/obsolete.html#the-marquee-element-0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
      <marquee id="test1" scrolldelay="a">Test1</marquee>
      <marquee id="test2" scrolldelay="-1">Test2</marquee>
      <marquee id="test3" scrolldelay="1">Test3</marquee>
      <marquee id="test4" scrolldelay="100">Test4</marquee>
    <script>
      setup({explicit_done:true});

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test1");
        assert_equals(mq.scrollDelay, 85, "The delay time should be 85ms.");
      }, "marquee_scrolldelay_string");}, 100);

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test2");
        assert_equals(mq.scrollDelay, 85, "The delay time should be 85ms.");
      }, "marquee_scrolldelay_less_than_zero");}, 100);

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test3");
        assert_equals(mq.scrollDelay, 60, "The delay time should be 60ms.");
      }, "marquee_scrolldelay_less_than_60");}, 100);

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test4");
        assert_equals(mq.scrollDelay, 100, "The delay time should be 100ms.");
      }, "marquee_scrolldelay_normal");}, 100);

      setTimeout(function(){done()}, 120);
    </script>
  </body>
</html>
